Feature/mosek direct api (#2)
* Add highs validation for quadratic constraints

* Multi-dimensional quadratic constraints - Fixed LP file and Gurobi direct API export to handle constraints with coordinates

* Add quadratic constraints to netcdf serialization

* add test_netcdf_roundtrip_multidimensional

* Added the following properties to MatrixAccessor in matrices.py:

  | Property  | Description                                             |
  |-----------|---------------------------------------------------------|
  | qclabels  | Vector of labels of all quadratic constraints           |
  | qc_sense  | Vector of senses (<=, >=, =)                            |
  | qc_rhs    | Vector of right-hand-side values                        |
  | Qc        | List of sparse Q matrices (one per constraint)          |
  | qc_linear | Sparse matrix of linear coefficients (n_qcons × n_vars) |

  The Q matrices follow the convention x'Qx where:
  - Diagonal terms are doubled (for x²)
  - Off-diagonal terms are symmetric (for xy)

* Fix imports

* Add dual support

* Add Add mosek direct api

* qc_linear property now handles the case when there are no linear terms in the quadratic constraints (the vars column doesn't exist in
   that case)

* 1. generatevarnames: Changed np.arange(0, len(labels)) to np.arange(0, len(labels), dtype=np.int32) - MOSEK expects 32-bit integers for indices
  2. putarowslice: Added explicit type casts:
    - indptr → np.int64 (pointer arrays)
    - indices → np.int32 (column indices)
    - data → np.float64 (values)

+++ linopy/io.py
@@ -672,7 +672,12 @@ def to_mosek(
 
     labels = np.vectorize(print_variable)(M.vlabels).astype(object)
     task.generatevarnames(
-        np.arange(0, len(labels)), "%0", [len(labels)], None, [0], list(labels)
+        np.arange(0, len(labels), dtype=np.int32),
+        "%0",
+        [len(labels)],
+        None,
+        [0],
+        list(labels),
     )
 
     ## Variables
@@ -731,7 +736,12 @@ def to_mosek(
         if M.A is not None:
             A = M.A.tocsr()
             task.putarowslice(
-                0, m.ncons, A.indptr[:-1], A.indptr[1:], A.indices, A.data
+                0,
+                m.ncons,
+                A.indptr[:-1].astype(np.int64),
+                A.indptr[1:].astype(np.int64),
+                A.indices.astype(np.int32),
+                A.data.astype(np.float64),
             )
             task.putconboundslice(0, m.ncons, bkc, blc, buc)
